fix(workouts): store precise pace values for single-file imports
Single-file imports were truncating pace values to integers while bulk imports stored precise decimal values. This created inconsistent data where the same workout would have different pace precision depending on import method.

Changes:
- Remove (int) cast from pace calculation in ProcessSingleFile
- Change CreateWorkoutEntity avgPaceS parameter from int to double
- Ensures consistency with Workout.AvgPaceS double precision field and bulk import behavior

+++ api/Endpoints/WorkoutsEndpoints.cs
@@ -2657,7 +2657,7 @@ public static class WorkoutsEndpoints
         DateTime startedAtUtc,
         int durationSeconds,
         double distanceMeters,
-        int avgPaceS,
+        double avgPaceS,
         double? elevationGainMeters,
         byte[] rawFileData,
         string fileName,
@@ -3161,7 +3161,7 @@ public static class WorkoutsEndpoints
 
             // Calculate average pace (seconds per km - stored in metric)
             var avgPaceS = distanceMeters > 0 && durationSeconds > 0
-                ? (int)(durationSeconds / (distanceMeters / 1000.0))
+                ? durationSeconds / (distanceMeters / 1000.0)
                 : 0;
 
             // Extract additional metrics from RawGpxData JSON (for GPX files)
